requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pydantic-settings
pydantic[email]
python-multipart
Pillow
firebase-admin
stripe
